requirements.txt
python-dotenv # For reading environment variables stored in .env file
langchain 
langchain-community 
langchain-openai # For embeddings
unstructured # Document loading
# onnxruntime==1.17.1 # chromadb dependency: on Mac use `conda install onnxruntime -c conda-forge`
# For Windows users, install Microsoft Visual C++ Build Tools first
# install onnxruntime before installing `chromadb`
chromadb # Vector storage
openai # For embeddings
tiktoken # For embeddings 
psutil
unstructured[md] # pip install "unstructured[md]" #do this for *.ms files
unstructured[pdf] # pip install "unstructured[pdf]" #do this for *.PDF files
pdfplumber 
markdownify
PyMuPDF 
pytesseract # Tesseract is an open source OCR or optical character recognition engine https://github.com/UB-Mannheim/tesseract/wiki
streamlit # fOR THE GUID
